requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1
faiss-cpu>=1.13.0
pypdf>=4.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0